requirements.txt
aiohttp==3.10.10
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-api-core==2.21.0
google-auth==2.35.0
google-generativeai==0.8.3
httplib2==0.22.0
python-dotenv==1.0.1
tenacity==9.0.0